♻️  Switch `producers/java-findsecbugs` to produce fileURI targets
